deprecate index for key prop and item soft-delete
Array index was used for the React key prop value for list rendering. The index is no longer used and was replaced by pseudo-unique id generated with Date.now() method.

Soft-delete of the tasks was deprecated since there was no usage for it. Additionally, the collection of tasks is kept in the local storage - limiting number of tasks will prevent issues related to reaching limit of local storage capacity. Soft-delete also caused issues with implementing grid layout for the task list since soft-deleted items weren't removed from the grid layout when not rendered.